Rewrite indentation support.
The indentation support has been moved to its own file.  This is an
attempt to modularize areas of the code containing a number of related functions.

The best-effort indentation support has been rewritten to simplify the
implementation in a more maintainable approach.  This resulted in a
number of general node-related functions which were placed into a new
"core" file, for utilization across other areas, such as completion.
Other common functions where also moved to this new file.

The tree-sitter indentation rules have been rewritten as tree-sitter
queries to eliminate many predicate functions.

GNAT Project keywords which are out-dented have been electrified.
When `electric-indent-mode` is enabled but `RET` is remapped (such as
to `newline-and-indent` in Doom Emacs), it's not guaranteed that the
current line will be re-indented when pressing `RET`.  As a result,
electric indentation is necessary to cause out-denting.