Make Eglot uniformly configurable.
Some Emacs distributions (e.g., Doom Emacs) expect to be able to
configure all Eglot modes and servers through the use of the standard
`eglot-server-programs`.  Therefore, `eglot-server-programs` is no
longer initialized buffer locally during mode initialization, and
instead the global version is configured for `gpr-ts-mode` (if
needed), when the Eglot package is loaded.  Earlier versions of
Eglot (e.g., the version shipped with Emacs 29) did not support
`gpr-ts-mode`.  Therefore, after `gpr-ts-mode` and `eglot` are loaded,
a check is performed to see if `gpr-ts-mode` is configured in
`eglot-server-programs`.  If not, a configuration is added.